DP_Lab2_110704071_彭叡楀/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
Pillow==10.2.0
tqdm==4.66.2
matplotlib==3.8.3
